Time for planets

+++ BlackHole.java
@@ -1,5 +1,5 @@
 class BlackHole extends MassObject {
-    public BlackHole(int mass, int age, String name, int[] position) {
+    public BlackHole(int mass, int age, String name, double[] position) {
         super(mass, age, name, position);
     }
 }
